requirements.txt
python-dotenv
vncdotool
pydantic
pymongo
boto3
